@@ -2,17 +2,16 @@
   "private": true,
   "name": "lisuying.com",
   "version": "0.0.0",
-  "license": "MIT",
   "homepage": "https://lisuying.com",
   "scripts": {
     "start": "per-env",
     "start:production": "npm run -s serve",
     "start:development": "npm run -s dev",
     "build": "preact build",
-    "serve": "preact build && preact serve",
+    "deploy": "preact build --no-prerender && cp ./src/CNAME build/ && gh-pages -d build",
+    "serve": "preact build && serve build --single",
     "dev": "preact watch",
     "lint": "eslint src",
-    "deploy": "preact build --no-prerender && cp ./src/CNAME build/ && gh-pages -d build",
     "test": "jest"
   },
   "eslintConfig": {
@@ -22,49 +21,31 @@
     "build/*"
   ],
   "devDependencies": {
-    "eslint": "^4.9.0",
-    "eslint-config-synacor": "^2.0.2",
+    "enzyme": "^3.10.0",
+    "enzyme-adapter-preact-pure": "^2.0.0",
+    "eslint": "^6.0.1",
+    "eslint-config-synacor": "^3.0.4",
     "identity-obj-proxy": "^3.0.0",
-    "jest": "^25.0.0",
-    "node-sass": "^4.12.0",
+    "jest": "^24.9.0",
+    "jest-preset-preact": "^1.0.0",
     "per-env": "^1.0.2",
-    "preact-cli": "^2.1.0",
+    "preact-cli": "^3.0.0-rc.6",
     "preact-render-spy": "^1.2.1",
-    "sass-loader": "^7.1.0"
+    "serve": "^11.1.0",
+    "node-sass": "^4.13.0",
+    "sass-loader": "^7.3.1"
   },
   "dependencies": {
     "gh-pages": "^2.0.1",
     "preact": "^10.0.1",
-    "preact-async-route": "^2.2.1",
-    "preact-router": "^2.5.7"
+    "preact-render-to-string": "^5.1.0",
+    "preact-router": "^3.0.0"
   },
   "jest": {
-    "verbose": true,
+    "preset": "jest-preset-preact",
     "setupFiles": [
-      "<rootDir>/tests/__mocks__/browserMocks.js"
-    ],
-    "testRegex": "(/(__tests__|tests)/.*|(\\.|/)(test|spec))\\.jsx?$",
-    "testPathIgnorePatterns": [
-      "/node_modules/",
-      "<rootDir>/tests/__mocks__/*"
-    ],
-    "testURL": "http://localhost:8080",
-    "moduleFileExtensions": [
-      "js",
-      "jsx"
-    ],
-    "moduleDirectories": [
-      "node_modules"
-    ],
-    "moduleNameMapper": {
-      "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/tests/__mocks__/fileMock.js",
-      "\\.(css|less|scss)$": "identity-obj-proxy",
-      "^./style$": "identity-obj-proxy",
-      "^preact$": "<rootDir>/node_modules/preact/dist/preact.min.js",
-      "^react$": "preact-compat",
-      "^react-dom$": "preact-compat",
-      "^create-react-class$": "preact-compat/lib/create-react-class",
-      "^react-addons-css-transition-group$": "preact-css-transition-group"
-    }
+      "<rootDir>/tests/__mocks__/browserMocks.js",
+      "<rootDir>/tests/__mocks__/setupTests.js"
+    ]
   }
 }
