@@ -1,53 +1,47 @@
-import { Component } from 'preact';
-import Data from '../lib';
+import { h } from 'preact';
+import { useState, useEffect } from 'preact/hooks';
 import { route } from 'preact-router';
 
-class Projects extends Component {
-    state = {
-        project: null
-    }
-    componentDidMount(){
-        if(this.props.matches.name && Data[this.props.matches.name]){
-            this.setState({
-                project: Data[this.props.matches.name]
-            })
-        }else{
+export default function Projects(props) {
+    const [project, setProject] = useState({})
+    const [content, setContent] = useState([])
+
+    useEffect(() => {
+        if(props.matches.projectName == "/"){
             setTimeout(() => {
                 route('/')
-            }, 2000)
+            }, 1000)
+        }else{
+            setProject(props.overview[props.matches.projectName])
+            setContent(props.content[props.matches.projectName])
         }
-    }
-    render() {
-        const project = this.state.project
-        return (
-            project
-            ?
-            <div className='project'>
-                <div className='header'>
-                    <div className='left'>
-                        <div className='header-title'>
-                            <h2>{project.context}</h2>
-                            <h1>{project.name}</h1>
-                        </div>
-                        <div className='header-img' style={{backgroundImage: `url(${project.cover})`}}/>
+    }, [])
+
+    return (
+        <div className='project container'>
+            <div className='header'>
+                <div className='header-left'>
+                    <div className='header-title'>
+                        <h2>{project.context}</h2>
+                        <h1>{project.name}</h1>
                     </div>
-                    <div className='right'>
-                        <div className='header-info-grid'>
-                            <h3>year</h3><h2>{project.year}</h2>
-                            <h3>tools</h3><h2>{project.tools}</h2>
-                            <h3>role</h3><h2>{project.role}</h2>
-                        </div>
-                    </div>
+                    <div className='header-img' style={{backgroundImage: `url(${project.cover})`}}/>
                 </div>
-                <div className='body'>
-                    {project.body}
+                <div className='header-right'>
+                    <div className='header-info'>
+                        <h3>year</h3><h2>{project.year}</h2>
+                        <h3>tools</h3><h2>{project.tools}</h2>
+                        <h3>role</h3><h2>{project.role}</h2>
+                    </div>
+                    {
+                        project.website &&
+                        <a href={project.website} target="_blank" className='underline'>visit website</a>
+                    }
                 </div>
             </div>
-            :
-            <div className='project container'>
-                <p>project not found, redirecting...</p>
+            <div className='body'>
+                    {content}
             </div>
-        );
-    }
+        </div>
+    );
 }
-export default Projects;
